BackEnd/Ai_counsellor/requirements.txt
django>=5.2.10
djangorestframework
django-cors-headers
djangorestframework-simplejwt
psycopg2-binary
python-dotenv
gunicorn
python-decouple
django-allauth
openai
langgraph
langchain
cryptography
google-generativeai